minor refactored and new source code org
